21 Sep 2009

+++ common/user_cmd.h
@@ -9,7 +9,7 @@
 typedef struct usercmd_s
 {
 	int		msec;
-	float		servertime;
+	long		servertime;
 	vec3_t		angles;
 	int		forwardmove;
 	int		sidemove;
